+++ frontend/config-overrides.js
@@ -1,0 +1,10 @@
+// module.exports = {
+//   webpack: function (config, env) {
+//     config.devServer = {
+//       ...config.devServer,
+//       path: "/dicom",
+//       target: "http://45.135.165.211:8181",
+//     };
+//     return config;
+//   },
+// };
